pyproject.toml
[project]
name = "mglyph-ml"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "albumentationsx>=2.0.12",
    "ipywidgets>=8.1.8", # temporarily here because of a mglyph bug
    "matplotlib>=3.10.7",
    "mglyph>=0.5.5",
    "nbstripout>=0.8.1",
    "notebook>=7.4.7",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "pytest>=8.4.2",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
]

[tool.uv]
package = true
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu130" },
  ],
]

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"

[project.optional-dependencies]
cpu = [
  "torch>=2.9.0",
  "torchvision>=0.24.0",
]
cu130 = [
  "torch>=2.9.0",
  "torchvision>=0.24.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu130", extra = "cu130" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu130", extra = "cu130" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://pypi.org/simple/"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
